Uncommented a code fragment, that was commented by mistake

+++ src/cuda/device/cuda_kernel_matrix_transform.cu
@@ -4,7 +4,7 @@
 
 #include "../../../include/mole_math/macros.h"
 
-__global__ void cuda_kernel_matrix_subtract_rows(float *row_minuend_values, float *row_subtrahend_values, double multiplier, size_t cols) {
+__global__ void cuda_kernel_matrix_subtract_rows(double *row_minuend_values, double *row_subtrahend_values, double multiplier, size_t cols) {
     size_t index = GLOBAL_IDX_X();
     size_t stride = GLOBAL_STRIDE_X();
