+++ src/cuda/device/cuda_kernel_matrix_transform.cuh
@@ -3,6 +3,6 @@
 
 const int threads_per_block = 512;
 
-__global__ void cuda_kernel_matrix_subtract_rows(float *row_minuend_values, float *row_subtrahend_values, double multiplier, size_t cols);
+__global__ void cuda_kernel_matrix_subtract_rows(double *row_minuend_values, double *row_subtrahend_values, double multiplier, size_t cols);
 
 #endif
